+++ packages/dropdown/src/react/item.tsx
@@ -1,3 +1,6 @@
+import ActionMenu from '@pluralsight/ps-design-system-actionmenu'
+import { CheckIcon } from '@pluralsight/ps-design-system-icon'
+import { HTMLPropsFor } from '@pluralsight/ps-design-system-util'
 import React, {
   forwardRef,
   useContext,
@@ -6,19 +9,16 @@ import React, {
   useImperativeHandle
 } from 'react'
 import { css } from 'glamor'
-import ActionMenu from '@pluralsight/ps-design-system-actionmenu'
-import { DropdownContext } from '../js'
-import { CheckIcon } from '@pluralsight/ps-design-system-icon'
-import { HTMLPropsFor } from '@pluralsight/ps-design-system-util'
 
 import stylesheet from '../css'
+import { DropdownContext } from '../js'
 
 const styles = {
   icon: css(stylesheet['.psds-dropdown--selected-icon']),
   text: css(stylesheet['.psds-dropdown--item-text'])
 }
 
-interface DropdownItemProps extends HTMLPropsFor<'li'> {
+interface DropdownItemProps extends Omit<HTMLPropsFor<'button'>, 'ref'> {
   icon?: React.ReactNode
   value?: React.ReactText
   menu?: React.ReactNode
@@ -29,7 +29,7 @@ export const Item = forwardRef<HTMLLIElement, DropdownItemProps>(
     const selectedValue = useContext(DropdownContext)
     const showSelectedValue = value && selectedValue === value
 
-    const ref = useRef<HTMLLIElement | null>(null)
+    const ref = useRef<HTMLLIElement>()
     useImperativeHandle(forwardedRef, () => ref.current)
 
     useEffect(() => {
@@ -38,15 +38,8 @@ export const Item = forwardRef<HTMLLIElement, DropdownItemProps>(
       }
     }, [showSelectedValue])
 
-    // TODO: update ActionMenu.Item ref typings then come back and remove casting
     return (
-      <ActionMenu.Item
-        ref={ref as any}
-        tagName="button"
-        value={value}
-        nested={menu}
-        {...rest}
-      >
+      <ActionMenu.Item tagName="button" value={value} nested={menu} {...rest}>
         <ActionMenu.Icon marginLeft>{icon}</ActionMenu.Icon>
         <ActionMenu.Ellipsis {...styles.text}>{children}</ActionMenu.Ellipsis>
         {showSelectedValue && <CheckIcon {...styles.icon} />}