+++ packages/dropdown/src/react/menu.tsx
@@ -7,7 +7,10 @@ import {
 import ActionMenu from '@pluralsight/ps-design-system-actionmenu'
 import stylesheet from '../css'
 
-const styles = css(stylesheet['.psds-dropdown__menu'])
+const styles = {
+  menuWrapper: css(stylesheet['.psds-dropdown__menu-wrapper']),
+  menu: css(stylesheet['.psds-dropdown__menu'])
+}
 
 interface DropdownMenuProps extends Omit<HTMLPropsFor<'ul'>, 'onClick'> {
   inNode?: HTMLElement
@@ -24,8 +27,9 @@ export const Menu = forwardRef<HTMLUListElement, DropdownMenuProps>(
       menu &&
       isOpen &&
       createUniversalPortal(
-        <div {...styles} style={menuPosition}>
+        <div {...styles.menuWrapper} style={menuPosition}>
           <ActionMenu
+            {...styles.menu}
             onClick={onClick}
             ref={ref}
             onClose={onClose}